🧪 cache plugin playground
Split files
