+++ apps/sandbox/src/app/app.module.ts
@@ -7,7 +7,7 @@ import { HttpExtModule } from '@http-ext/http-ext';
 import { AppComponent } from './app.component';
 import { BookDetailComponent } from './book-detail.component';
 import { BookListComponent } from './book-list.component';
-import { cachePlugin } from './http/cache-plugin';
+import { cachePlugin } from './http/cache-plugin/cache-plugin';
 import { loggerPlugin } from './http/logger-plugin';
 
 export const routes: Routes = [
