+++ apps/sandbox/src/app/http/cache-plugin/providers/indexedDB-provider.ts
@@ -1,0 +1,11 @@
+import { CacheProvider } from './provider';
+
+export class IndexedDBProvider implements CacheProvider {
+  get(url) {
+    throw new Error('todo');
+  }
+
+  set(url, response) {
+    throw new Error('todo');
+  }
+}
