+++ apps/sandbox/src/app/http/cache-plugin/providers/memory-provider.ts
@@ -1,0 +1,13 @@
+import { CacheProvider } from './provider';
+
+export class MemoryProvider implements CacheProvider {
+  private cache = new Map<string, unknown>();
+
+  get(url) {
+    return this.cache.get(url);
+  }
+
+  set(url, response) {
+    this.cache.set(url, response);
+  }
+}
