+++ apps/sandbox/src/app/http/cache-plugin/providers/provider.ts
@@ -1,0 +1,6 @@
+export type CacheProviderType = 'memory' | 'localStorage' | 'indexedDB';
+
+export interface CacheProvider {
+  get(url: string): unknown;
+  set(url: string, response: unknown): void;
+}
